AV-214603 Moving HTTPRoute Validation to status layer
